+++ src/main/java/io/github/binout/soccer/domain/notification/TemplateEngine.java
@@ -1,0 +1,8 @@
+package io.github.binout.soccer.domain.notification;
+
+import java.util.Map;
+
+public interface TemplateEngine {
+    
+    String render(String templateName, Map<String, Object> params);
+}
